Rework the "List Of Forum Helpers" page to use JSON (#20)
-This update is focused on reworking the "List Of Forum Helpers" page in order to make it use JSON to display the list. This is better than the old system because it will be easier to add new Forum Helpers or modify existing Forum Helpers.

This was inspired by #15 but it has been reworked to use JavaScript instead of Jekyll.

Note: Biographies/users might not be up to date. This is intentional, because the JSON files were copied from #15 which was created several months ago. These biographies/missing users will be fixed in a later update after this pull request has been merged