Add Contributors Page (#87)
Resolves #36

Change Log:

Add contributors page sorted in a grid fashion. It is sorted from most to least commits. For the people who contributed outside of commits (eg. issue creation), it is sorted alphabetically. It is generated using JavaScript, so if someone else contributes, their GitHub username can be added to the contributors array and they will be added to the page.
Added a link to the contributors page and GitHub repository to the footer.
Made labeler label PRs involving the contributors page as "documentation".
Next Steps:

Add more information to the page (eg. what kind of contributions did the person make?)
Add badges to the list of Forum Helpers page indicating if someone has contributed to the site